Add reference to next/prev in use collider
Previously choosing a higher maxColliders number would result in
significantly worse performance, because all colliders had to be
iterated and then checked whether they were in use.

My first approach to solve this is by saving an index to the next and
previous in-use collider, so I can iterate the collider array like a
doubly-linked list. The bookkeeping for these indices is simple in most
cases, *except* for collider creation. The code is not very simple,
which makes me consider trying another approach.

Performance Numbers:
Previously: maxColliders =  2000 => ~0.07ms
            maxColliders = 20000 => ~0.45ms (oof!)

Now:        maxColliders =  2000 => <0.04ms
            maxColliders = 20000 => <0.05ms

This does solve the goal of having performance be less dependent on
maxColliders and also makes it faster in general.

One very interesting (odd) thing to note is that the frame time starts
out very low and rises for a good number of seconds before it converges.
It does converge though!