Use a free range instead of free list
Just store a maxInUseIndex and insert new colliders after it, swap out
destroyed colliders.
No colliders are skipped, only inUse colliders are iterated over and
they are iterated in-order. Bookkeeping is essentially nothing.

For some reason this is still a bit slower than the doubly-linked-list,
but faster than the free list.